requirements.txt
python-dotenv>=1.0.1
pydantic>=2.7.0
langchain>=0.2.11
langchain-openai>=0.1.7
openai>=1.40.0
python-pptx>=0.6.23
requests>=2.31.0
streamlit>=1.39.0
